requirements.txt
python-can==4.6.1
cantools==41.0.2
pyserial==3.5
flask==3.0.2
pyusb==1.3.1; platform_system == "Windows"
gs-usb==0.3.0; platform_system == "Windows"
libusb-package==1.0.26.3; platform_system == "Windows"
